Make CCMBridge.PerClassSingleNodeCluster extensible.
This commit changes the signature of configure() so that it is possible to instantiate and return subclasses of Cluster.Builder.
It also suppresses the jvmArgs static field that imposes a specific query handler for all tests and replaces it with a new method withJvmArgs() in CCMBridge.Builder, configurable on a per-test basis.