Set idempotence on Statement instead of globally
Makes test more reliable as its possible any ongoing queries could be
executed and increment the speculative execution counter.

Also reattempt test if scheduled execution did not occur.  It is
possible, yet unlikely, that a scheduled execution may not fire if the
timeout is not invoked before the response is received.   This should be
incredibly rare.