Wait up to 10 seconds for keyspace presence
It is not guaranteed that the token map is updated when refreshSchema
completes.  The reason why is that the node list refresh is submitted,
but the future completion is not tied to the node list completing.  This
is fixed by JAVA-1193, so this commit is not neccessary when that is
merged.